% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_BBS.R
\name{buffer_BBS}
\alias{buffer_BBS}
\title{buffer_BBS}
\usage{
buffer_BBS(alpha, bbs, buffer = 2, method = "rec", raw = FALSE)
}
\arguments{
\item{buffer}{Numeric value indicating the number of degrees to buffer the count data (default = 2)}

\item{method}{Which buffer method to use? Options include: "rec", "bcr", "bcr + rec" (default = "rec")}

\item{spp_count}{Data frame containing the count data for the focal species obtained using the `GetSppCounts()` function}

\item{routes}{Data frame containing the route data obtained from the function `GetRoutes()`}
}
\value{
A .csv file containing the following fields:

routeID The unique 8 digit route ID for each route

Year The year that the count was conducted

aou The numeric code for the focal species

countN Total individuals of the focal species recorded at stop N (50-stop data) or N-9:N (10-stop data)

Latitude The latitude for the route

Longitude The longitude for the route
}
\description{
Add all routes within rectangular buffer around the maximum and minimum observed latitude and longitude
}

