% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_eval.R
\name{drop.bad.coef}
\alias{drop.bad.coef}
\title{drop.bad.coef}
\usage{
drop.bad.coef(keep, cov)
}
\arguments{
\item{mod}{Model output file}
}
\value{
TRUE if model converged; FALSE if not
}
\description{
Drops quadratic term if linear term is bad
}

