% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMods.R
\name{GetGamMods}
\alias{GetGamMods}
\title{GetGamMods}
\usage{
GetGamMods()
}
\value{
A list containing the model covariates for each of the 961 potential gamma/epsilon models
}
\description{
Returns list with full model set for 961 gamma/epsilon models
}

