% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIndices.R
\name{GetIndices}
\alias{GetIndices}
\title{GetIndices}
\usage{
GetIndices(prob_df)
}
\arguments{
\item{prob_df}{Data frame containing the annual occupancy probabilities (from GetOcc)}
}
\value{
Data frame containing annual estimates of the following indices:

avg.occ = range-wide mean occupancy probability

s.lat = southern range limit

n.lat = northern range limit

avg.lat = occupancy-weighted mean breeding latitude

avg.lon = occupancy-weighted mean breeding longitude
}
\description{
Estimate indices of range dynamics from annual occupancy estimates
}

