% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof}
\alias{gof}
\title{gof}
\usage{
gof(alpha, pao)
}
\arguments{
\item{alpha}{Alpha code for species of interest}

\item{pao}{.pao file}
}
\description{
Identify top model, test GOF, save output, and delete output files of all other models
}

