% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{input_data}
\alias{input_data}
\title{input_data}
\usage{
input_data(year_range = seq(from = 1972, to = 2014), bbs_data = bbs,
  clim_data = NA_biovars)
}
\description{
Formats BBS data, climate data, and presence input files for all species in spp_list
}

