% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_route_plot.R
\name{make_route_plot}
\alias{make_route_plot}
\title{make_route_plot}
\usage{
make_route_plot(buff_routes, used_routes, raw_routes, alpha)
}
\description{
plot all routes, including counts, buffer, and outliers
}

