% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSummary.R
\name{MakeBetatab}
\alias{MakeBetatab}
\title{MakeBetatab}
\usage{
MakeBetatab(coefs, sd.err, alpha, covs_use, nYears, years, nuisance = FALSE)
}
\arguments{
\item{nuisance}{If true, returns AIC table of theta, theta', p, & omega; if false, returns AIC table for psi, gamma & epsilon}
}
\description{
Make Beta table for including in report
}

