% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_eval.R
\name{mod_eval}
\alias{mod_eval}
\title{mod_eval}
\usage{
mod_eval(pres_out, mod, pao2, dig = 3, large = 8, time, het)
}
\arguments{
\item{mod}{Model output file}

\item{dig}{Minimum threshold for significant digits}

\item{large}{Maximum threshold for beta coefficients}
}
\value{
TRUE if model converged; FALSE if not
}
\description{
Checks for problems with convergence
}

