% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glob_opts.R
\name{model_opts}
\alias{model_opts}
\title{model_opts}
\usage{
model_opts(psi.test = FALSE, gam.test = FALSE, het = TRUE,
  annual = TRUE, Parallel = TRUE, limit.cores = 50)
}
\description{
Set model options for running presence
}

