% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_BBSclim.R
\name{run_BBSclim}
\alias{run_BBSclim}
\title{run_BBSclim}
\usage{
run_BBSclim(annual, runmods, ...)
}
\arguments{
\item{annual}{TRUE/FALSE Should annual models be fit for all species in spp_list_annual?}

\item{runmods}{TRUE/FALSE Should psi & gamma/epsilon models be fit for all species in spp_list_runmods?}

\item{...}{Arguments to be passed to model_opts()}
}
\description{
Run full analysis
}

