% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMods.R
\name{RunPsiMods}
\alias{RunPsiMods}
\title{RunPsiMods}
\usage{
RunPsiMods(alpha)
}
\arguments{
\item{alpha}{alpha code for species of interest}
}
\description{
Runs full model set for 127 psi models, evalute each, write and return AIC table, delete output file (optional)
}

