% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{sim.bbs.ms}
\alias{sim.bbs.ms}
\title{sim.bbs.ms}
\usage{
sim.bbs.ms(psi.coefs = psi.coefs, th0.coefs = th0.coefs,
  th1.coefs = th1.coefs, gam.coefs = gam.coefs, eps.coefs = eps.coefs,
  p1.coefs = p1.coefs, p2.coefs = p2.coefs, pi.coefs = pi.coefs,
  years = years)
}
\description{
Simulate BBS data using covariates and coefficients from top model
}

