% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_het.R
\name{test_het}
\alias{test_het}
\title{test_het}
\usage{
test_het(alpha, pao)
}
\arguments{
\item{alpha}{Alpha code for species of interest}

\item{pao}{pao file}
}
\value{
AIC table comparing model with and without heterogeneity in p
}
\description{
Tests whether there is support for heterogeneous p
}

