% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMods.R
\name{top_covs}
\alias{top_covs}
\title{top_covs}
\usage{
top_covs(alpha, mods, psi = TRUE)
}
\arguments{
\item{mods}{List of models}

\item{psi}{Return top psi model (TRUE) or gamma/epsilon model (FALSE)}

\item{aic_tab}{AIC table from RunPsiMods or RunGamMods}
}
\description{
Get covariates of top model
}

