% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzipCRU.R
\name{unzipCRU}
\alias{unzipCRU}
\title{Unzip CRU climate data.}
\usage{
unzipCRU(varb, in_dir = "data-raw/", out_dir = "data-raw/")
}
\arguments{
\item{varb}{The climate variable of interest}

\item{in_dir}{The directory containing the compressed files}

\item{out_dir}{The director where the uncompressed files will be stored}
}
\description{
\code{unzipCRU} unzips the compressed CRU climate data and
adds the decompressed files to the "data" folder for each variable.
}

