% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPres.R
\name{write_dm_and_run2}
\alias{write_dm_and_run2}
\title{write_dm_and_run2}
\usage{
write_dm_and_run2(pao, cov_list, out, het, time, dm_list = NULL, modname,
  fixed = FALSE, inits = FALSE, maxfn = 32000, alpha = NULL,
  parse = FALSE)
}
\arguments{
\item{pao}{.pao file for analysis}

\item{cov_list}{List containing the covariates for each parameter}

\item{out}{Which output folder should results be written to?}

\item{dm_list}{List containing the design matrices for each parameter}

\item{modname}{Optional name for .out file with model results}

\item{fixed}{vector of fixed values (eg., fixed(1)=0, fixed(4)=.1 -> fixed=c(0,.1); names(fixed)=c(1,4))}

\item{maxfn}{maximum number if iterations for optimization routine}

\item{initvals}{vector of beta initial values}
}
\value{
A .out file containing the fitted model and parameters estimates (see ?occ.mod for details)
}
\description{
Write design matrices to file and run correlated detection model in Presence. Note that this function calls Presence in a slightly different way than the `write_dm_and_run()` function in the current version of RPresence
}

