% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pao.R
\name{write_pao}
\alias{write_pao}
\title{write_pao}
\usage{
write_pao(counts, covs, alpha, sim = FALSE, is.tenstops)
}
\arguments{
\item{counts}{Data frame containing the buffered BBS counts obtained from the function `buffer_BBS()`}

\item{alpha}{Four letter alpha code for the species of interest}

\item{clim}{Data frame containing the annual bioclim values obtained from the function 'GetBioVars()`}
}
\value{
A .pao file containing the detection histories, covariates, and summary information to input into Presence
}
\description{
Write occupancy and climate data for program Presence
}

