% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMods.R
\name{GetGamMods}
\alias{GetGamMods}
\title{GetGamMods}
\usage{
GetGamMods(psi_covs)
}
\arguments{
\item{psi_cov}{Vector containing the climate covariates from the top psi model}
}
\value{
A list containing the model covariates for each of the 961 potential psi models
}
\description{
Returns list with full model set for 961 gamma/epsilon models
}

