% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMods.R
\name{GetPsiMods}
\alias{GetPsiMods}
\title{GetPsiMods}
\usage{
GetPsiMods()
}
\value{
A list containing the model covariates for each of the 31 potential psi models
}
\description{
Returns list with full model set for 31 psi models
}

