% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof}
\alias{gof}
\title{gof}
\usage{
gof(aic_tab, mods, cov_data, year_seq, stops, ...)
}
\arguments{
\item{aic_tab}{AIC table from RunGamMods}

\item{mods}{List of models}

\item{cov_data}{Data frame containing the covariates used to fit the full model}

\item{year_seq}{Vector containing the years included in the analysis}
}
\description{
Identify top model, test GOF, save output, and delete output files of all other models
}

