% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\name{input_data}
\alias{input_data}
\title{input_data}
\usage{
input_data(start_yr = 1972, end_yr = 2014)
}
\description{
Formats BBS data, climate data, and presence input files for all species in spp_list
}

