% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_route_plot.R
\name{make_route_plot}
\alias{make_route_plot}
\title{make_route_plot}
\usage{
make_route_plot(spp_buff, spp_counts, spp_counts2, alpha)
}
\description{
plot all routes, including counts, buffer, and outliers
}

