% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMods.R
\name{RunGamMods}
\alias{RunGamMods}
\title{RunGamMods}
\usage{
RunGamMods(alpha, pao)
}
\arguments{
\item{alpha}{four letter species code}

\item{pao}{pao object}
}
\description{
Runs full model set for 961 gamma/epsilon models, evaluate each, write and return AIC table, delete .out files (optional)
}

