% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMods.R
\name{RunGamMods}
\alias{RunGamMods}
\title{RunGamMods}
\usage{
RunGamMods(pao, alpha, mods = gam_mods, del = TRUE, ..., test = FALSE,
  trim = TRUE, Parallel = TRUE)
}
\arguments{
\item{gam_mods}{list containing the covariates for each model}
}
\description{
Runs full model set for 961 gamma/epsilon models, evaluate each, write and return AIC table, delete .out files (optional)
}

