% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMods.R
\name{RunPsiMods}
\alias{RunPsiMods}
\title{RunPsiMods}
\usage{
RunPsiMods(alpha, pao, mods = psi_mods)
}
\arguments{
\item{alpha}{alpha code for species of interest}

\item{pao}{.pao file for species of interest}

\item{psi_mods}{list containing the parameters for each model to evaluate}
}
\description{
Runs full model set for 31 psi models, evalute each, write and return AIC table, delete output file (optional)
}

