% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{sim.bbs.ms}
\alias{sim.bbs.ms}
\title{sim.bbs.ms}
\usage{
sim.bbs.ms(covs, psi.coefs, th0.coefs, th1.coefs, gam.coefs, eps.coefs,
  p1.coefs, p2.coefs, pi.coefs, years, cov_data, time, het, pao)
}
\description{
Simulate BBS data using covariates and coefficients from top model
}

