% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{test.presence.gof}
\alias{test.presence.gof}
\title{test.presence.gof}
\usage{
test.presence.gof(modname, large = 4, pao2, mod, is.annual, is.het, alpha,
  Psi.coefs, Gam.coefs, Eps.coefs, Psi.se, Gam.se, Eps.se)
}
\value{
1 if simulated and observed estimates are similar (model not overfit); 0 otherwise
}
\description{
Bootstrap GOF test: test if parameters from top model run with simulated data are consistent with original estimates
}

