% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_annual.R
\name{test_annual}
\alias{test_annual}
\title{test_annual}
\usage{
test_annual()
}
\value{
AIC table comparing model with and without annual variation in p
}
\description{
Tests whether there is support for annual variation in p
}

