% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_annual.R
\name{test_annual}
\alias{test_annual}
\title{test_annual}
\usage{
test_annual(alpha, pao)
}
\arguments{
\item{alpha}{Alpha code for species of interest}
}
\value{
TRUE if p should vary annual; otherwise FALSE
}
\description{
Tests whether there is support for annual variation in p
}

